requirements.txt
pydantic>=2.7
rich>=13.7
typer>=0.12
pandas>=2.2
python-dotenv>=1.0
ollama>=0.1.0
python-dotenv>=1.0.0
